Updated checks on ItemController

+++ backend/ShoppingList/src/main/java/org/adriantd/shoppinglist/lists/controller/ItemController.java
@@ -22,23 +22,37 @@ public class ItemController {
     private final CurrentUserService currentUserService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<List<ItemResponse>> getListItems(@PathVariable int id) {
+    public ResponseEntity<List<ItemResponse>> getListItems(@PathVariable(required = false) int id) {
+        if(id <= 0){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
         return ResponseEntity.ok(itemService.getAllItemsFromListId(id));
     }
 
     @PostMapping("/add")
-    public ResponseEntity<ItemResponse> addItem(@RequestBody RegisterItemRequest registerItemRequest) {
+    public ResponseEntity<ItemResponse> addItem(@RequestBody(required = false) RegisterItemRequest registerItemRequest) {
+        if (!isValidRegisterItemRequest(registerItemRequest)) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
         try{
             return ResponseEntity.ok(itemService.addItemToList(registerItemRequest,currentUserService.getCurrentUserNickname()));
         } catch (AccessDeniedException e) {
+            //Error on permissions
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         } catch (Exception e) {
+            //Any entity not found
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 
     @PostMapping("/remove")
-    public ResponseEntity<Void> removeItem(@RequestBody ItemRequest itemRequest) {
+    public ResponseEntity<Void> removeItem(@RequestBody(required = false) ItemRequest itemRequest) {
+        if (!isValidItemRequest(itemRequest)) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
         try{
             itemService.removeItemsFromRequest(itemRequest,currentUserService.getCurrentUserNickname());
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
@@ -51,9 +65,10 @@ public class ItemController {
 
     @PutMapping("/state")
     public ResponseEntity<Void> updateItemPurchasedState(@RequestBody(required = false) ItemRequest itemRequest) {
-        if(itemRequest == null){
+        if(!isValidItemRequest(itemRequest)){
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
+
         try{
             itemService.updateItemsPurchased(itemRequest, currentUserService.getCurrentUserNickname());
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
@@ -64,4 +79,15 @@ public class ItemController {
         }
     }
 
+    private boolean isValidItemRequest(ItemRequest itemRequest){
+        return itemRequest != null && itemRequest.getShoplistId() != null && itemRequest.getProductIds() != null;
+    }
+
+    private boolean isValidRegisterItemRequest(RegisterItemRequest registerItemRequest){
+        return registerItemRequest != null
+                && registerItemRequest.getShoplistId() != null
+                && registerItemRequest.getProductId() != null
+                && registerItemRequest.getUnits() != null
+                && registerItemRequest.getType() != null;
+    }
 }
